server/pyproject.toml
[project]
name = "rule_of_three"
version = "0.1.0"
description = "The rule of three application"
requires-python = "==3.11.9"
dependencies = [
    "pyyaml==6.0.1",
    "azure-cosmos==4.7.0",
    "azure-identity==1.17.1",
    "dataclass-wizard==0.22.3",
    "tenacity==9.0.0",
    "flask==3.0.3",
    "azure-functions==1.20.0",
    "flask-cors>=5.0.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "flask-caching>=2.3.0",
]

[tool.uv]
dev-dependencies = [
    "colorama >=0.4.6,<0.5.0",
    "iniconfig >=2.0.0,<3.0.0",
    "packaging >=23.2,<24.0",
    "pluggy >=1.4.0,<2.0.0",
    "pytest >=8.0.0,<9.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "flake8 >=7.0.0,<8.0.0",
    "black >=24.3.0,<25.0.0",
    "jwcrypto>=1.5.6",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
testpaths = [
  "tests"
]
pythonpath = [
  ".", "src",
]
